Updated various functions to use int64_t instead of int for contig coordinates.  Specifically anything using the total contig length or coordinates in the contig selector.
This means that the contig selector now displays data when the total
contig length is > 2Gb.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1715 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/gap5/contig_selector.h
@@ -34,7 +34,7 @@ typedef struct {
 int
 find_position_in_DB(GapIO *io, 
 		 int c_num, 
-		 int position);
+		 int64_t position);
 
 void
 PlotRepeats(GapIO *io, 
@@ -58,7 +58,7 @@ update_contig_order(Tcl_Interp *interp,
 		    int cs_id,
 		    int *order_array,
 		    int num_contigs,
-		    int cx);
+		    int64_t cx);
 
 void
 ReOrder(GapIO *io,
